+++ dropshot/examples/pagination-multiple-sorts.rs
@@ -237,7 +237,7 @@ async fn example_list_projects(
     query: Query<PaginationParams<ProjectScanParams, ProjectScanPageSelector>>,
 ) -> Result<HttpResponseOk<ResultsPage<Project>>, HttpError> {
     let pag_params = query.into_inner();
-    let limit = rqctx.page_limit(&pag_params)?.get();
+    let limit = rqctx.page_limit(&pag_params)?.get() as usize;
     let data = rqctx.context();
     let scan_params = ProjectScanParams {
         sort: match &pag_params.page {
